requirements.txt
fastmcp>=2.0.0
google-cloud-secret-manager==2.20.0
google-cloud-aiplatform==1.70.0
google-cloud-logging==3.5.0
vertexai>=1.0.0
supabase==2.9.1
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1
ccxt==4.4.23
python-telegram-bot==21.7
telethon==1.37.0
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.7
uvicorn==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
websockets==13.1
apscheduler==3.10.4
loguru==0.7.2
pandas_ta_classic==0.1.5
matplotlib==3.9.3
mplfinance==0.12.10b0
Pillow==11.0.0
pytz==2024.2
starlette==0.41.3
sse-starlette==2.1.3
